Track wallclock reception times of messages more carefully.
If the mlat aggregation timeout is set high, we might have a lot of
e.g. DF11 messages spanning a few seconds. If we track the approximate
arrival time of each message, we can get a better estimate of the time
that our position fix is valid at.

Also, when we're receiving from a Radarcape, we can directly find the
reception time from the GPS timestamp, so use that where available.