Use a new UUID field internally to uniquely identify a receiver connection.
The "user" of a receiver is still tracked separately (and is used when e.g.
reporting positions), but client connnection implementations can also provide
a UUID which might be different and which is used for anything that must be
unique per connection. The multiple-login rule is relaxed to only restrict
multiple logins with the same UUID.

(cherry picked from commit dd4a0899df9090eff46401132cb9167a4c12d4c9)